requirements.txt
python-telegram-bot==20.7
selenium
pdfplumber
pdf2image
pillow
flask
